chore: adjust the file directory organization structure
(cherry picked from commit 0c75f501ba55a2338380c12c18af0c49dc6d7d13)
