+++ components/utils/noop.ts
@@ -1,0 +1,1 @@
+export default function noop() {}
